requirements.txt
beautifulsoup4==4.3.2
cached-property==1.4.0
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11
django-cors-headers==2.2.0
django-crispy-forms==1.4.0
django-durationfield==0.5.1
django-filter==1.1.0
django-toolkit==0.3.3
django-url-filter==0.3.4
djangorestframework==3.7.7
enum-compat==0.0.2
enum34==1.1.6
gunicorn==19.8.1
idna==2.6
libsass==0.14.1
lxml==3.4.4
Markdown==2.6.11
mysqlclient==1.3.12
psycopg2==2.7.4
python-dateutil==2.4.2
pytz==2018.3
requests==2.18.4
six==1.11.0
tldextract==1.6
unicodecsv==0.13.0
urllib3==1.22
whitenoise==3.3.1
